% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiGET.R
\name{apiGET}
\alias{apiGET}
\title{apiGET - Do a get to the Core REST API.}
\usage{
apiGET(coreApi, resource, query, headers = NULL, special = NULL,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{resource}{entity type to get}

\item{query}{query string}

\item{headers}{headers to be added to get.}

\item{special}{- passed to buildUrl for special sdk endpoints}

\item{useVerbose}{Use verbose communication for debugging}

\item{encode}{encoding to use for request option are "multipart", "form", "json", "raw"}
}
\value{
RETURN return the entire http response
}
\description{
\code{apiGET}   Do a get to the Core REST API.
}
\details{
\code{apiGET}  Base call to Core REST API.
}
\examples{
\dontrun{
api<-CoreAPIV2::CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
response <-CoreAPIV2::apiGET(login$coreApi,"json",resource,query)
logOut(login$coreApi )
}
}
\author{
Craig Parman
}

