% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiPOST.R
\name{apiPOST}
\alias{apiPOST}
\title{apiPOST - Do a POST to the Core REST API.}
\usage{
apiPOST(coreApi, body = NULL, encode, special = NULL, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{body}{body for request}

\item{encode}{encoding to use for request option are "multipart", "form", "json", "raw"}

\item{special}{- passed to buildUrl for special sdk endpoints}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN return the entire http response
}
\description{
\code{apiPOST}  Base call to Core REST API.
}
\details{
\code{apiPOST}  BDo a POST to the Core REST API.
}
\examples{
\dontrun{
api<-CoreAPIV2::CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
response <-CoreAPIV2::apiPOST(login$coreApi,body,"json",,special=NULL,useVerbose=FALSE)
logOut(login$coreApi )
}
}
\author{
Craig Parman
}

