% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributecleanName.R
\name{attributeCleanName}
\alias{attributeCleanName}
\title{attributeCleanName - converts names to uppercase and replaces spaces with underscores to 
reference attribute names. 
\code{attributeCleanName} clean attribute name.}
\usage{
attributeCleanName(name)
}
\arguments{
\item{name}{string to clean}
}
\value{
Returns name in ODAt compliant form
}
\description{
\code{attriureCleanName} - onverts names to uppercase and replaces spaces with underscores to 
reference attribute names.
}
\examples{
\dontrun{
new_name <-CoreAPIV2::attributeCleanName("My Comments")
new_name
MY_COMMENTS
 }
}
\author{
Craig Parman, ngsAnalytics, ngsanalytics.com
}

