% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildJsonUrl.R
\name{buildJsonUrl}
\alias{buildJsonUrl}
\title{buildJsonUrl - build URL for call to Core REST API.}
\usage{
buildJsonUrl(coreApi, special = NULL, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{special}{flag for special sdk endpoints}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN Core json REST URL
}
\description{
\code{buildJsonUrl} build URL for call to Core json REST API.
}
\details{
\code{buildJsonUrl}  buils url for Core json REST API.
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
response <-CoreAPIV2::apiCall(login$coreApi,body,"json",useVerbose=FALSE)
logOut(login$coreApi )
}
}
\author{
Craig Parman
}

