% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildUrl.R
\name{buildUrl}
\alias{buildUrl}
\title{buildUrl - build URL for call to Core ODATA API.}
\usage{
buildUrl(coreApi, resource = NULL, query = NULL, special = NULL,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{resource}{resource path (required except for special requests)}

\item{query}{and additional property options (optional)}

\item{special}{flag for special sdk endpoints}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN Core REST URL
}
\description{
\code{buildUrl} build URL for call to Core REST API.
}
\details{
\code{buildUrl}  build URL for call to Core ODATA API.
}
\examples{
\dontrun{
api<-CoreAPIV2("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
response <-CoreAPIV2::buildUrl(coeApi,"Sample","('PS1')")
logOut(login$coreApi )
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com
}

