% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEntity.R
\name{createEntity}
\alias{createEntity}
\title{createEntity - Create a new instance of a entity. Used for standard object or "Plan Old Core Objects" (POCOs)}
\usage{
createEntity(coreApi, entityType, body, locationId = NULL,
  projectIds = NULL, barcode = NULL, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{entity type to get as character string}

\item{body}{values for attributes and associations as a  list of key-values pairs}

\item{locationId}{location ID for initial location as character string}

\item{projectIds}{project comma separated list of project IDs as character string}

\item{barcode}{User provided barcode as a character string}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{createEntity} Creates a new entity instance. Required inputs are url, jsessionId and entityType.
}
\details{
\code{createEntity} Create a new instance of a entity. Used for standard object or "Plan Old Core Objects" (POCOs)
}
\examples{
\dontrun{
api<-CoreAPIV2::coreApi("PATH TO JSON FILE")
login<- CoreAPI::authBasic(api)
item<-CoreAPI::createEntity(login$coreApi,"Entity_Type")
CoreAPI::logOut(login$coreApi )
}
}
\author{
Craig Parman
}

