% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExperiment.R
\name{createExperiment}
\alias{createExperiment}
\title{createExperiment - Create a new instance of an experiment.}
\usage{
createExperiment(coreApi, experimentType, assayType, assayBarcode, protocolType,
  protocolBarcode, body = NULL, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentType}{experiment type to get as character string}

\item{assayType}{assay type}

\item{assayBarcode}{assay barcode}

\item{protocolType}{protocol type}

\item{protocolBarcode}{protocol barcode}

\item{body}{values for experiment attributes and associations as a  list of key-values pairs}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{createExperiment} Creates a new experiment.
}
\details{
\code{createExperiment} Creates a new instance of an entity.
}
\examples{
\dontrun{
api<-CoreAPIV2::coreApi("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
experiment<-CoreAPIV2::createExperiment(login$coreApi,"Experiment_Type","Assaybarcode","Protocolbarcode")
CoreAPIV2::logOut(login$coreApi )
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com
}

