% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSampleLot.R
\name{createSampleLot}
\alias{createSampleLot}
\title{createSampleLot - Creates a lot of a sample.}
\usage{
createSampleLot(coreApi, sampleType, sampleBarcode, body = NULL,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{sampleType}{sample type to create the lot of}

\item{sampleBarcode}{parent sample barcode}

\item{body}{attributes as list of key-values pairs (optional)}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{createSampleLot} Creates a new sample lot using the parent sample barcode
}
\details{
\code{createSampleLot} Creates a sample lot of a sample. Assumes the sample lot entity is inth from samplename_lot
}
\examples{
\dontrun{
api<-CoreAPIV2("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
item<-CoreAPIV2::createSampleLot(login$coreApi,"Sample_Name")
logOut(login$coreApi )
}
}
\author{
Craig Parman
}

