% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentPublish.R
\name{experimentPublish}
\alias{experimentPublish}
\title{experimentPublish Publishes an experiment.}
\usage{
experimentPublish(coreApi, experimentType, experimentBarcode,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentType}{experiment entity type}

\item{experimentBarcode}{barcode of the experiment}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains updated experiment information, $response contains the entire http response
}
\description{
\code{experimentPublish} - Publishes an experiment.
}
\details{
\code{experimentPublish} Publishes an experiment.
}
\examples{
\dontrun{
api<-CoreAPIV2("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
update<- CoreAPIV2::experimentPublish(login$coreApi,experimentType, exptbarcode,useVerbose = TRUE)
CoreAPIV2::logOut(login$coreApi )
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com
}
