% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentUnPublish.R
\name{experimentUnPublish}
\alias{experimentUnPublish}
\title{experimentUnPublish UnPublishes an experiment.}
\usage{
experimentUnPublish(coreApi, experimentType, experimentBarcode,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentType}{experiment entity type}

\item{experimentBarcode}{barcode of the experiment}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains updated experiment information, $response contains the entire http response
}
\description{
\code{experimentUnPublish} - UnPublishes an experiment.
}
\details{
\code{experimentUnPublish} UnPublishes an experiment.
}
\examples{
\dontrun{
api<-CoreAPIV2("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
update<- CoreAPIV2::experimentUnPublish(login$coreApi,experimentType, exptbarcode,useVerbose = TRUE)
CoreAPIV2::logOut(login$coreApi )
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com
}

