% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellContents.R
\name{getCellContents}
\alias{getCellContents}
\title{getCellContents -  Gets information about container cell contents.}
\usage{
getCellContents(coreApi, cellId, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{cellId}{Cell ID returned in $ID slot of contaner cell, not to be confused with the content ID}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains contents information.  This is list with the cell contents 
in cell number order. $response contains the entire http response.
}
\description{
\code{getCellContents} - Gets information about container cell contents.
}
\details{
\code{getCellContents} Gets information about container cell contents.
}
\examples{
\dontrun{
api<-CoreAPIV2::coreApi("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
cell<-CoreAPIV2::getCellContents(login$coreApi,cellID,useVerbose = TRUE)
CoreAPIV2::logOut(login$coreApi )
}
}
\author{
Craig Parman
}

