% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContainerContents.R
\name{getContainerContents}
\alias{getContainerContents}
\title{getContainerContents -  Gets information about container cell contents.}
\usage{
getContainerContents(coreApi, containerType, containerBarcode,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{containerType}{Container type}

\item{containerBarcode}{container barcode}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains contents information.  This is list with the cell contents 
in cell number order. $response contains the entire http response.
}
\description{
\code{getCellContents} - Gets information about container cell contents.
}
\details{
\code{getContainerlContents} Gets information about container cell contents.
}
\examples{
\dontrun{
api<-CoreAPIV2::coreApi("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
cell<-CoreAPIV2::getContainerContents(login$coreApi,"VIAL","VIA9")
CoreAPIV2::logOut(login$coreApi )
}
}
\author{
Craig Parman
}

