% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEntityAssociations.R
\name{getEntityAssociation}
\alias{getEntityAssociation}
\title{\code{getEntityAssociation} Get assoication for a context}
\usage{
getEntityAssociation(coreApi, entityType, barcode, context,
  fullMetadata = TRUE, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{entity type to get}

\item{barcode}{barcode of entity to get}

\item{context}{association context}

\item{fullMetadata}{- get full metadata}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http POST}
}
\value{
returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{getEntityAssociation}  Get assoication for a context
}
\examples{
\dontrun{
api<-CoreAPIV2::CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
associations<-CoreAPIV2::getEntityAssociation(login$coreApi,"entityType","barcode",values)
CoreAPIV2::logOut(login$coreApi)
}
}
\author{
Craig Parman
}

