% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEntityByName.R
\name{getEntityByName}
\alias{getEntityByName}
\title{getEntityByName - Get an entity by name from the Core LIMS using the ODATA API.}
\usage{
getEntityByName(coreApi, entityType, name, fullMetadata = TRUE,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{entity type to get}

\item{name}{name of entity to get}

\item{fullMetadata}{- get full metadata}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http POST}
}
\value{
returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{getEntityByName}  Get an entity by barcode from the Core LIMS using the ODATA API.
}
\details{
\code{getEntityByName} get an entity from the LIMS by barcode
}
\examples{
\dontrun{
api<-CoreAPIV2::CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
item<-CoreAPIV2::getEntityByName(login$coreApi,"entityType","name")
CoreAPIV2::logOut(login$coreApi)
}
}
\author{
Craig Parman
}

