% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEntityMetadata.R
\name{getEntityMetadata}
\alias{getEntityMetadata}
\title{getEntitMetadata - Get entity metadata.}
\usage{
getEntityMetadata(coreApi, entityType, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{entity type to get metadata for}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http POST}
}
\value{
returns a list containing $attributes for attribute properties, 
and $association for association properties; $template contains a list that can be converted to JSON for object creation.
}
\description{
\code{getEntityMetadata} Get an entity metadata by entityType. Returns a list with three data frames
named attributes,associations, and template. Template can be used as the body to for createEntity functions.
}
\details{
\code{getEntityMetadata}  Get entity metadata.
}
\examples{
\dontrun{
api<-CoreAPIV2::CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
item<-CoreAPIV2::getEntityMetadata(login$coreApi,"entityType","barcode")
CoreAPIV2::logOut(login$coreApi)
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com
}
