% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentContainers.R
\name{getExperimentContainers}
\alias{getExperimentContainers}
\title{getExperimentContainers - Gets experiment samples from experiment identified by barcode.}
\usage{
getExperimentContainers(coreApi, experimentType, barcode, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentType}{experiment entity type to get}

\item{barcode}{barcode of entity to get}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http POST}
}
\value{
returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{getExperimentContainers}  Gets experiment contaiers from experiment identified by experiment barcode. 
Gets samples attached as sample lots and in containers.
}
\details{
\code{getExperimentContainers}  Gets experiment samples from experiment identified by barcode.
}
\examples{
\dontrun{
api<-CoreAPIV2::CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
experiment<-  CoreAPIV2::getExperimentContainers(login$coreApi,"entityType","barcode")
CoreAPIV2:logOut(login$coreApi)
}
}
\author{
Craig Parman
}

