% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentSamples.R
\name{getExperimentSamples}
\alias{getExperimentSamples}
\title{getExperimentSamples - Gets experiment sample barcodes from experiment identified by barcode.}
\usage{
getExperimentSamples(coreApi, experimentType, barcode, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentType}{experiment entity type to get}

\item{barcode}{barcode of entity to get}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http POST}
}
\value{
returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{ getExperimentSamples}  Gets experiment sample barcodes from experiment identified by experiment barcode.
}
\details{
\code{getExperimentSamples}  Gets experiment samples from experiment identified by barcode.
}
\examples{
\dontrun{
api<-CoreAPIV2::CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
response<-  CoreAPIV2::getExperimentSamples(login$coreApi,"entityType","barcode")
experimentsampleBarcodes <- response$entity
CoreAPIV2:logOut(login$coreApi)
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com
}

