% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentSampleAssayFileData.R
\name{getExperimentSamplesAssayFileData}
\alias{getExperimentSamplesAssayFileData}
\title{getExperimentSamplesAssayFileData - Gets file attached as assay data.}
\usage{
getExperimentSamplesAssayFileData(coreApi, assayType, experimentSamplebarcode,
  attributeName, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{assayType}{assay type to get}

\item{experimentSamplebarcode}{experiment sample barcode of entity to get}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http POST}

\item{attrributeName}{Name of the attibute that containts the file data}
}
\value{
returns a list $entity containsbinary object that in the file, $response contains the entire http response
}
\description{
\code{ getExperimentSamplesAssayFileData }  Gets file attached as assay data. 
Use getExperimentSamplesAssayData for non-file data.
}
\details{
\code{getExperimentSamplesAssayFileData }  Gets file attached as assay data.
}
\examples{
\dontrun{
api<-CoreAPIV2::CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
response<-  getExperimentSamplesAssayFileData(login$coreApi,"assayType","barcode","CI_FILE")
writeBin(response$entity, "myfile.png")
CoreAPIV2:logOut(login$coreApi)
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com
}

