% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentSampleIntermediateData.R
\name{getExperimentSamplesIntermediateData}
\alias{getExperimentSamplesIntermediateData}
\title{getExperimentSamplesIntermediateData - Gets intermediate data for an experiment sample.}
\usage{
getExperimentSamplesIntermediateData(coreApi, experimentType, assayType,
  dataName, experimentSamplebarcode, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentType}{experiment type for sample}

\item{assayType}{assay type for sample}

\item{dataName}{assay data name to retrive as configured in the assay.}

\item{experimentSamplebarcode}{experiment sample barcode of entity to get}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http call}
}
\value{
returns a list $entity contains data frame with derived experiment sample barcodes concentration, 
        and assay raw data. $response contains the entire http response
}
\description{
\code{getExperimentSamplesIntermediateData}   Gets raw data for a experiment sample identified by barcode.
}
\details{
\code{getExperimentSamplesIntermediateData}  Gets intermediate data for a experiment sample identified by barcode.
}
\examples{
\dontrun{
api<-CoreAPIV2::CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
response<-  getExperimentSamplesIntermediateData(login$coreApi,"ExperimentType","assayType","dataName","barcode")
rawdata <- response$entity
CoreAPIV2:logOut(login$coreApi)
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com
}

