% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWellContents.R
\name{getWellContents}
\alias{getWellContents}
\title{getWellContents -  Gets information about a single container well contents.}
\usage{
getWellContents(coreApi, containerType, containerBarcode, containerWellNum,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{containerType}{container type}

\item{containerBarcode}{container barcode}

\item{containerWellNum}{container well number as a string}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains cell information, $response contains the entire http response
}
\description{
\code{getWellContents} -  Gets information about a single container well contents.
}
\details{
\code{getWellContents} Gets information about container well contents.
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
cell<-CoreAPIV2::getWellContents(login$coreApi,"VIA9","1",fullMetadata = TRUE)
CoreAPIV2::logOut(login$coreApi )
}
}
\author{
Craig Parman
}

