% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSONapiCall.R
\name{JSONapiCall}
\alias{JSONapiCall}
\title{JSONapiCall - Base call to Core REST API.}
\usage{
JSONapiCall(coreApi, body, encode, headers = NULL, special = NULL,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{body}{body for request}

\item{encode}{encoding to use for request option are "multipart", "form", "json", "raw"}

\item{headers}{headers to add}

\item{special}{- passed to buildUrl for special sdk endpoints}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN return the entire http response
}
\description{
\code{JSONapiCall}  Base call to Core json REST API.
}
\details{
\code{JSONapiCall}  Base call to Core json REST API.
}
\examples{
\dontrun{
api<-CoreAPIV2::CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
response <-CoreAPIV2::JSONapiCall(login$coreApi,body,encode="json",,special=NULL,useVerbose=FALSE)
logOut(login$coreApi )
}
}
\author{
Craig Parman
}

