% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSONbuildUrl.R
\name{JSONbuildUrl}
\alias{JSONbuildUrl}
\title{JSONbuildUrl - build URL for call to Core REST API.}
\usage{
JSONbuildUrl(coreApi, special = NULL, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{special}{flag for special sdk endpoints}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN Core REST URL
}
\description{
\code{buildUrl} build URL for call to Core REST API.
}
\details{
\code{apiCall}  base call to Core REST API.
}
\examples{
\dontrun{
api<-COREAPIV2::CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
response <-CoreAPIV2::JSONapiCall(login$coreApi,body,"json",useVerbose=FALSE)
logOut(login$coreApi )
}
}
\author{
Craig Parman
}

