% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCellContents.R
\name{setCellContents}
\alias{setCellContents}
\title{setCellContents -  Puts a sample lot in a container well.}
\usage{
setCellContents(coreApi, containerType, containerBarcode, containerCellId,
  sampleLotType, sampleLotBarcode, amount, amountUnit, concentration,
  concentrationUnit, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{containerType}{container type}

\item{containerBarcode}{container barcode}

\item{containerCellId}{container cell id}

\item{sampleLotType}{sample lot type}

\item{sampleLotBarcode}{barcode of lot to add to cell}

\item{amount}{amount to add (numeric)}

\item{amountUnit}{units}

\item{concentration}{(numeric)}

\item{concentrationUnit}{concentration units}

\item{useVerbose}{use verbose communications for debugging}
}
\value{
RETURN returns a list $entity contains updated container
        information, $response contains the entire http response
}
\description{
\code{setCellContents} - Puts a cell lot in a container cell.
}
\details{
\code{setCellContents} Puts a sample lot in a container well.
}
\examples{
\dontrun{
api<-CoreAPIV2::CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
cell<- updateCellContents<-(coreApi, containerType,containerBarcode, containerCellId,
                           sampleLotType,sampleLotBarcode, amount, amountUnit, concentration,
                           concentrationUnit,useVerbose = FALSE)
CoreAPIV2::logOut(login$coreApi )
}
}
\author{
Craig Parman
}

