% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setExperimentSampleAssayFileData.R
\name{setExperimentSamplesAssayFileData}
\alias{setExperimentSamplesAssayFileData}
\title{setExperimentSamplesAssayFileData - puts file attached as assay data in an experiment.}
\usage{
setExperimentSamplesAssayFileData(coreApi, assayType, experimentSamplebarcode,
  attributeName, filepath, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{assayType}{assay type to get}

\item{experimentSamplebarcode}{experiment sample barcode of entity to get}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http POST}

\item{attrributeName}{Name of the attibute that containts the file data}

\item{filePath}{path to file to upload}
}
\value{
returns a list $entity containsbinary object that in the file, $response contains the entire http response
}
\description{
\code{ setExperimentSamplesAssayFileData } Puts file attached as assay data in an experiment
}
\details{
\code{setExperimentSamplesAssayFileData }  puts file attached as assay data.
}
\examples{
\dontrun{
api<-CoreAPIV2::CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
response<-  setExperimentSamplesAssayFileData(login$coreApi,"assayType","barcode","filepath","CI_FILE")
CoreAPIV2:logOut(login$coreApi)
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com
}

