% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setWellContents.R
\name{setWellContents}
\alias{setWellContents}
\title{setWellContents -  Puts a sample lot in a container well.}
\usage{
setWellContents(coreApi, containerBarcode, containerCellNum, sampleLotType,
  sampleLotBarcode, amount, amountUnit, concentration, concentrationUnit,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{containerBarcode}{container barcode}

\item{sampleLotType}{sample lot type}

\item{sampleLotBarcode}{barcode of lot to add to cell}

\item{amount}{amount to add (numeric)}

\item{amountUnit}{units}

\item{concentration}{(numeric)}

\item{concentrationUnit}{concentration units}

\item{useVerbose}{use verbose communications for debugging}

\item{containerWellNum}{container well number}
}
\value{
RETURN returns a list $entity contains updated container
        information, $response contains the entire http response
}
\description{
\code{setWellContents} - Puts a cell lot in a container cell.
}
\details{
\code{setWellContents} Puts a sample lot in a container well.
}
\examples{
\dontrun{
api<-CoreAPIV2::CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
cell<- updateCellContents<-(coreApi, containerBarcode, containerCellNum,
                           sampleLotType,sampleLotBarcode, amount, amountUnit, concentration,
                           concentrationUnit,useVerbose = FALSE)
CoreAPIV2::logOut(login$coreApi )
}
}
\author{
Craig Parman
}

