% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateExperimentSampleData.R
\name{updateExperimentSampleData}
\alias{updateExperimentSampleData}
\title{updateExperimentSampleData - Update experiment sample assay data.}
\usage{
updateExperimentSampleData(coreApi, assayType, experimentSamplebarcode,
  assayAttributeValues, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{assayType}{entity type to get as character string}

\item{experimentSamplebarcode}{User provided barcode as a character string}

\item{assayAttributeValues}{assay attributes as a list of key-values pairs}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains entity information,
       $response contains the entire http response
}
\description{
\code{updateExperimentSampleData} Update experiment sample assay data.
}
\details{
\code{updateExperimentSampleData} Update experiment sample assay data.
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
newdata<-CoreAPIV2::updateExperimentSampleData(login$coreApi,assayType,
                  experimentSampleBarcode,assayAtributeValues)
CoreAPIV2::logOut(login$coreApi ) response<- CoreAPI::authBasic(coreApi)
}
}
\author{
Craig Parman
}

