% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMetadata.R
\name{updateMetadata}
\alias{updateMetadata}
\title{\code{updateMetadata} g Updates cached metadata so metadata is up to date.}
\usage{
updateMetadata(coreApi, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http POST}
}
\value{
returns XML with all entitiy metadata
}
\description{
\code{updateMetadata}  Updates cached metadata so metadata is up to date. 
Must be run after any configuration changes.
}
\examples{
\dontrun{
api<-CoreAPIV2::CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
metadata<-CoreAPIV2::updateMetadata(login$coreApi,useverbose=TRUE)
CoreAPIV2::logOut(login$coreApi)
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com
}
