% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellContents.R
\name{getCellContents}
\alias{getCellContents}
\title{getCellContents -  Gets information about a single container well contents.}
\usage{
getCellContents(coreApi, containerType, containerBarcode, containerCellId,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{containerType}{container type}

\item{containerBarcode}{container barcode}

\item{containerCellId}{container cell number as a string}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains cell information, $response contains the entire http response
}
\description{
\code{getCellContents} -  Gets information about a single container cell contents.
}
\details{
\code{getCellContents} Gets information about container well contents.
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
cell<-CoreAPIV2::getCellContents(login$coreApi,"VIA9","1234233",fullMetadata = TRUE)
CoreAPIV2::logOut(login$coreApi )
}
}
\author{
Craig Parman
}

