% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContainerContents.R
\name{getContainerContents}
\alias{getContainerContents}
\title{getContainerContents -  Gets information about container cell contents.}
\usage{
getContainerContents(coreApi, containerType, containerBarcode,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{containerType}{container entity type}

\item{containerBarcode}{container barcode}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains cell information, $response contains the entire http response
}
\description{
\code{getContainerContents} - Gets information about container cell contents. This call uses the JSON API.
}
\details{
\code{getContainerContents} Gets information about container cell contents.
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
cell<-CoreAPIV2::ggetContainerContents(login$coreApi,"VIA9","1")
CoreAPIV2::logOut(login$coreApi )
}
}
\author{
Craig Parman
}

