% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentContainers.R
\name{getExperimentContainers}
\alias{getExperimentContainers}
\title{getExperimentContainers - Gets experiment containers from experiment identified by barcode.}
\usage{
getExperimentContainers(coreApi, experimentType, barcode, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentType}{experiment entity type to get}

\item{barcode}{barcode of experiment to query}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http POST}
}
\value{
returns a list $entity contains barcodes of the containers, $response contains the entire http response
}
\description{
\code{getExperimentContainers}  Gets experiment contaniers from experiment identified by experiment barcode.
}
\details{
\code{getExperimentContainers}  Gets experiment containers from experiment identified by barcode.
}
\examples{
\dontrun{
api<-CoreAPIV2::CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
exptCaontainerBarcodes <-  CoreAPIV2::getExperimentContainers(login$coreApi,"entityType","barcode")
CoreAPIV2:logOut(login$coreApi)
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com
}

