% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentSampleAssayData.R
\name{getExperimentSamplesAssayData}
\alias{getExperimentSamplesAssayData}
\title{getExperimentSamplesAssayData - Gets assay data for an experiment sample.}
\usage{
getExperimentSamplesAssayData(coreApi, assayType, experimentSamplebarcode,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{assayType}{assay type to get}

\item{experimentSamplebarcode}{experiment sample barcode of entity to get}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http POST}
}
\value{
returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{ getExperimentSamplesAssayData }  Gets experiment samples from experiment identified by experiment barcode. 
Does not retieve files attached as data. Use getExperimentSampleAssayFileData to retrieve assay data that is a file.
}
\details{
\code{getExperimentSamplesAssayData }  Gets assay data for a experiment sample identified by barcode.
}
\examples{
\dontrun{
api<-CoreAPIV2::CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
experiment<-  getExperimentSamplesAssayData (login$coreApi,"assayType","barcode")
CoreAPIV2:logOut(login$coreApi)
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com
}
