% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateEntityAssociations.R
\name{updateEntityAssociations}
\alias{updateEntityAssociations}
\title{\code{updateEntityAssociations} Update entity assoications}
\usage{
updateEntityAssociations(coreApi, entityType, barcode, updateValues,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{entity type to get}

\item{barcode}{barcode of entity to get}

\item{updateValues}{values to update as list of associations contex and entity type pair and barcode}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http POST}
}
\value{
returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{updateEntityAssociations}  Update entity associations.
}
\examples{
\dontrun{
api<-CoreAPIV2::CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
item<-CoreAPIV2::updateEntityAssociations(login$coreApi,"entityType","barcode",values)
CoreAPIV2::logOut(login$coreApi)
}
}
\author{
Craig Parman
}

