% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateEntityLocation.R
\name{updateEntityLocation}
\alias{updateEntityLocation}
\title{\code{updateEntityLocation}- Update entity location}
\usage{
updateEntityLocation(coreApi, entityType, barcode, locationBarcode,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{entity type to get}

\item{barcode}{barcode of entity to get}

\item{locationBarcode}{loaction barcode}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http POST}
}
\value{
returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{updateEntityLocation} - Update entity location
}
\examples{
\dontrun{
api<-CoreAPIV2::CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
response<-CoreAPIV2::updateEntityLocation(login$coreApi,"entityType","barcode","locationBarcode")
entity <- response$entity
CoreAPIV2::logOut(login$coreApi)
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com
}
